refactor(agent-ux): split 509-line agent-ux.ts into 4 focused modules
agent-ux.ts mixed four largely-independent concerns: error
classification, source parsers, refactor helpers, and clause/fixity
inference. Split into siblings under `src/agda/`:

  error-classifier.ts  (188) — TriageClass / TriageResult, classifier
                               branches, rewriteCompilerPlaceholders,
                               extractSuggestedRename, normalizeConfidence.
  source-parsers.ts    (171) — OPTIONS pragma, .agda-lib flags, module
                               shape, top-level defs, postulate sites.
  refactor-helpers.ts  (131) — splitWords, matchesTypePattern,
                               applyScopedRename, buildAutoSearchPayload,
                               plus ScopedRenameResult / AutoSearchOptions.
  clause-fixity.ts     (159) — inferMissingClauseArity, buildMissingClause,
                               inferFixityConflicts plus FixityConflict
                               and the curated DEFAULT_IMPORTED_FIXITIES.
  agent-ux.ts           (71) — barrel re-exporting everything so
                               existing imports across `tools/agent-ux/*`
                               keep working unchanged.

splitWords lives in refactor-helpers and is consumed by source-parsers
and clause-fixity, so the dependency graph is a clean DAG (no
circular imports). Each new file is well under the 500-line ceiling.

No behavior change: 1186 vitest cases pass identically.
